user role

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -50,6 +50,10 @@ class CategoryController extends Controller
      */
     public function store(StoreCategoryRequest $request)
     {
+
+        if (auth()->user()->role !== 'admin')
+            abort(404);
+
         Category::create($request->all());
         return back()->with('success-message', 'Category created successfully !');
     }
@@ -64,6 +68,10 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
+
+        if (auth()->user()->role !== 'admin')
+            abort(404);
+
         $search = request('search');
         $categories = Category::where('name', 'like', '%' . $search . '%')
             ->orderBy('id', 'DESC')
@@ -93,6 +101,10 @@ class CategoryController extends Controller
      */
     public function update(UpdateCategoryRequest $request, Category $category)
     {
+
+        if (auth()->user()->role !== 'admin')
+            abort(404);
+
         $category->update($request->all());
         return back()->with('success-message', 'Category updated successfully !');
     }
@@ -105,6 +117,11 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+
+
+        if (auth()->user()->role !== 'admin')
+            abort(404);
+
         $category->delete();
         return redirect()->back()->with('success-message2', 'Category deleted successfully');
     }
